working with edit reservations

+++ app/_lib/actions.js
@@ -4,6 +4,7 @@ import { revalidatePath } from 'next/cache';
 import { auth, signIn, signOut } from './auth';
 import supabase from './supabase';
 import { getBookings } from './data-service';
+import { redirect } from 'next/navigation';
 
 export async function signInAction() {
   await signIn('google', { redirectTo: '/account' });
@@ -44,6 +45,9 @@ export async function deleteReservation(bookingId) {
   const guestBookings = await getBookings(session.user.guestId);
   const guestBookingsIds = guestBookings.map((booking) => booking.id);
 
+  if (!guestBookingsIds.includes(bookingId))
+    throw new Error('You are not allowed to delete this booking!');
+
   const { error } = await supabase
     .from('bookings')
     .delete()
@@ -54,4 +58,37 @@ export async function deleteReservation(bookingId) {
   }
 
   revalidatePath('/account/reservations');
+}
+
+export async function updateBooking(formData) {
+  const session = await auth();
+  const bookingId = Number(formData.get('bookingId'));
+  if (!session) throw new Error('You must be logged in');
+
+  const guestBookings = await getBookings(session.user.guestId);
+  const guestBookingsIds = guestBookings.map((booking) => booking.id);
+
+  if (!guestBookingsIds.includes(bookingId))
+    throw new Error('You are not allowed to edit this booking!');
+
+  const updateData = {
+    numGuests: Number(formData.get('numGuests')),
+    observations: formData.get('observations').slice(0, 1000),
+  };
+
+  const { error } = await supabase
+    .from('bookings')
+    .update(updateData)
+    .eq('id', bookingId)
+    .select()
+    .single();
+
+  if (error) {
+    throw new Error('Booking could not be updated');
+  }
+
+  revalidatePath(`/account/reservations/edit/${bookingId}`);
+  revalidatePath('/account/reservations');
+
+  redirect('/account/reservations');
 }
